+++ backend/tests/util.py
@@ -201,6 +201,7 @@ async def load_sample_structures(minimal=False):
             'org_unit_ean': 'e34d4426-9845-4c72-b31e-709be85d6fa2',
             'org_unit_email': '73360db1-bad3-4167-ac73-8d827c0c8751',
             'org_unit_telefon': '1d1d3711-5af4-4084-99b3-df2b8752fdec',
+            'org_unit_contactopenhours': 'e8ea1a09-d3d4-4203-bfe9-d9a2da100f3b',
             # org_unit_level
             'org_unit_level_10': '0f015b67-f250-43bb-9160-043ec19fad48',
             # visibility